# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GymFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List subscriptions
      parameters:
        - in: query
          name: memberId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [active, expired, pending]
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"

    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create a subscription for a member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get a subscription by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update a subscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionBody"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # ── Payments ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        planId:
          type: integer
        planName:
          type: ["string", "null"]
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, expired, pending]
        gracePeriodDays:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - memberId
        - planId
        - startDate
        - endDate
        - status
        - gracePeriodDays
        - createdAt
    CreateSubscriptionBody:
      type: object
      properties:
        memberId:
          type: integer
        planId:
          type: integer
        startDate:
          type: string
        gracePeriodDays:
          type: integer
      required:
        - memberId
        - planId
        - startDate
    UpdateSubscriptionBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, expired, pending]
        endDate:
          type: string
        gracePeriodDays:
          type: integer

    # ── Payments ──────────────────────────────────────────────────────────
...